backend/requirements.txt
Flask
Flask-Cors
requests
groq
langchain
langchain-groq
langchain-community
langchain-huggingface
langchain-chroma
chromadb
sentence-transformers
pypdf
python-dotenv
flask-sqlalchemy
flask-jwt-extended
psycopg2-binary
langchain-text-splitters
langchain-classic
tqdm
gunicorn